# api/swagger.yml
openapi: 3.0.0
info:
  description: |
    This API server is the layer that allows 3rd party apps and services interfacing with the coordinator to explore, monitor and use the Hermez rollup.
    Example of these apps are:
    * Wallet: send L2 transactions, check balance, ...
    * Explorer: List transactions, slots, batches, ...
    * Exchange integrations
    
    ### Pagination
    
    #### Usage

    All the endpoints that return a list of undefined size use pagination. Unless the opposite is explicitly said.
    In order to use pagination, three query parameters are used:
    * `fromItem`: indicates the first item to be returned. In general, this parameter shouldn't be provided in the first call to the endpoint, and use the `itemId` of the last returned item (+/-) 1, if the order is (ascending/descending).
    * `order`: all pginated items are ordered chronologicaly. However the specific fields to guarantee this order depend on each endpoint. For this purpose, `itemId` is used (itemId follows ascending chronological order except for unforged L1 user transactions). If the parameter is not provided, ascending order will be used by default. 
    * `limit`: maximum amount of items to include in each response. Default is 20, maximum 2049.
    
...

    #### Reorgs and safetyness

    Since all the items are ordered chronologicaly, there are no safety problems when fetching items in ascending order, except for reorgs (more on this later).
...

    ### Signatures

...

  version: "0.0.1"
...
servers:
...
paths:
...
  '/accounts':
    get:
      tags:
        - Account
      summary: Get accounts balances and other associated information.
      description: Get accounts balances and other associated information.
      operationId: getAccounts
      parameters:
        - name: hezEthereumAddress
          in: query
          description: Only get accounts associated to an Ethereum address. Incompatible with the query `BJJ`.
          required: false
          schema:
            $ref: '#/components/schemas/HezEthereumAddress'
        - name: BJJ
          in: query
          description: Only get accounts associated to a BabyJubJub public key.  Incompatible with the query `hezEthereumAddress`.
          required: false
          schema:
            $ref: '#/components/schemas/BJJ'
        - name: tokenIds
          in: query
          required: false
          description: Only get accounts of specific tokens.
          schema:
            type: string
            description: Comma separated list of token identifiers.
            example: "3,87,91"
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. Accounts will be ordered by increasing account index.
          schema:
            type: string
            default: ASC
            enum:
                - ASC
                - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accounts'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/accounts/{accountIndex}':
    get:
      tags:
        - Account
      summary: Get an account by its index.
      description: Get an account by its index.
      operationId: getAccount
      parameters:
        - name: accountIndex
          in: path
          description: Identifier of an account.
          required: true
          schema:
            $ref: '#/components/schemas/AccountIndex'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/exits':
    get:
      tags:
        - Account
      summary: Get exit information. This information is required to perform a withdraw.
      description: Get exit information. This information is required to perform a withdraw.
      operationId: getExits
      parameters:
        - name: tokenId
          in: query
          required: false
          description: Only get exits of specific token
          schema:
            $ref: '#/components/schemas/TokenId'
        - name: hezEthereumAddress
          in: query
          description: Get exits associated to a Ethereum address. Incompatible with query `BJJ` and `accountIndex`.
          required: false
          schema:
            $ref: '#/components/schemas/HezEthereumAddress'
        - name: BJJ
          in: query
          description: Get exits associated to a BabyJubJub public key. Incompatible with query `hezEthereumAddress` and `accountIndex`.
          required: false
          schema:
            $ref: '#/components/schemas/BJJ'
        - name: accountIndex
          in: query
          description: Get exits associated to a specific account. Incompatible with queries `tokenId`, `hezEthereumAddress` and `BJJ`.
          required: false
          schema:
            $ref: '#/components/schemas/AccountIndex'
        - name: batchNum
          in: query
          description: Get exits from the exit tree of a specific batch.
          required: false
          schema:
            $ref: '#/components/schemas/BatchNum'
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. Exits will be ordered by increasing (batchNum, accountIndex).
          schema:
            type: string
            default: ASC
            enum:
                - ASC
                - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exits'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/exits/{batchNum}/{accountIndex}':
    get:
      tags:
        - Account
      summary: Get specific exit information.
      description: Get exit information form a specific exit tree and account. This information is required to perform a withdraw. Exits are identified with accounIndex and batchNum since every batch that has exits has a different exit tree.
      operationId: getExit
      parameters:
        - name: batchNum
          in: path
          description: Batch of the exit tree.
          required: true
          schema:
            $ref: '#/components/schemas/BatchNum'
        - name: accountIndex
          in: path
          description: Account identifier.
          required: true
          schema:
            $ref: '#/components/schemas/AccountIndex'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exit'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
  '/transactions-pool/{id}':
    get:
      tags:
        - Transaction
      summary: Get details and status of a transaction that is in the pool.
      description: >-
        Get transaction from the pool by its id. This endpoint is specially useful for tracking the status of a transaction that may not be forged yet.
        Only transactions from the pool will be returned.
        Note that the transaction pool is different for each coordinator and therefore only a coordinator that has received a specific transaction
        will be able to provide information about that transaction.
      operationId: getPoolTx
      parameters:
        - name: id
          in: path
          description: Transaction identifier.
          required: true
          schema:
            $ref: '#/components/schemas/TransactionId'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolL2Transaction'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/transactions-history':
    get:
      tags:
        - Transaction
      summary: Get details and status of transactions that have been forged.
      description: >-
        Get historical transactions. This endpoint will return all the different types of transactions except for:
          - Transactions that are still in the transaction pool of any coordinator. These transactions can be fetched using `GET /transactions-pool/{id}`.
          - L1 transactions sent by users that have not been forged yet. These transactions can be fetched using `GET /transactions-history/{id}`.
      operationId: getHistoryTxs
      parameters:
        - name: tokenId
          in: query
          required: false
          description: Only get transactions of specific token
          schema:
            $ref: '#/components/schemas/TokenId'
        - name: hezEthereumAddress
          in: query
          required: false
          description: Only get transactions sent from or to an account associated to an Ethereum address Incompatible with the queries `BJJ` and `accountIndex`.
          schema:
            $ref: '#/components/schemas/HezEthereumAddress'
        - name: BJJ
          in: query
          description: Only get transactions associated to a BabyJubJub public key. Incompatible with the queries `hezEthereumAddress` and `accountIndex`.
          required: false
          schema:
            $ref: '#/components/schemas/BJJ'
        - name: accountIndex
          in: query
          required: false
          description: Only get transactions sent from or to a specific account. Incompatible with the queries `tokenId`, `hezEthereumAddress` and `BJJ`.
          schema:
            $ref: '#/components/schemas/AccountIndex'
        - name: batchNum
          in: query
          required: false
          description: Only get transactions forged in a specific batch.
          schema:
            $ref: '#/components/schemas/BatchNum'
        - name: type
          in: query
          required: false
          description: Only get transactions of a specific type.
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. History transactions will be ordered by (batchNum, position).
          schema:
            type: string
            default: ASC
            enum:
                - ASC
                - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryTransactions'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/transactions-history/{id}':
    get:
      tags:
        - Transaction
      summary: Get details and status of a historical transaction.
      description: >-
        Get transaction from the history by its id. This endpoint will return all the different types of transactions except those that are still in the pool of any coordinator.
      operationId: getHistoryTx
      parameters:
        - name: id
          in: path
          description: Transaction identifier.
          required: true
          schema:
            $ref: '#/components/schemas/TransactionId'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryTransaction'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/batches':
    get:
      tags:
        - Hermez status
      summary: Get information about forged batches.
      description: >-
        Get information about forged batches.
      operationId: getBatches
      parameters:
        - name: minBatchNum
          in: query
          required: false
          description: Include only `batchNum < minBatchNum` batches.
          schema:
            $ref: '#/components/schemas/BatchNum'
        - name: maxBatchNum
          in: query
          required: false
          description: Include only `batchNum > maxBatchNum` batches.
          schema:
            type: number
        - name: slotNum
          in: query
          required: false
          description: Include only batches that were forged within the specified slot.
          schema:
            $ref: '#/components/schemas/SlotNum'
        - name: forgerAddr
          in: query
          required: false
          description: Include only batches forged by `forgerAddr`
          schema:
            $ref: '#/components/schemas/EthereumAddress'
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. Batches will be ordered by increasing `batchNum`.
          schema:
            type: string
            default: ASC
            enum:
                - ASC
                - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batches'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'                
  '/batches/{batchNum}':
    get:
      tags:
        - Hermez status
      summary: Get a specific batch.
      description: >-
        Get a specific batch.
      operationId: getBatch
      parameters:
        - name: batchNum
          in: path
          description: Batch identifier.
          required: true
          schema:
            $ref: '#/components/schemas/BatchNum'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/full-batches/{batchNum}':
    get:
      tags:
        - Hermez status
      summary: Get a full batch
      description: >-
        Get a specific batch, including the associated transactions. The object returned in this method can be a bit heavy. 
        If you're devloping a front end, you may consider using a combinaton of `GET /batches/{batchnum}` and `GET /history-transactions?batchNum={batchNum}`.
      operationId: getFullBatch
      parameters:
        - name: batchNum
          in: path
          description: Batch identifier
          required: true
          schema:
            $ref: '#/components/schemas/BatchNum'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullBatch'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/slots':
    get:
      tags:
        - Hermez status
      summary: Get information about slots.
      description: >-
        Get information about slots.
      operationId: getSlots
      parameters:
        - name: minSlotNum
          in: query
          required: false
          description: Only include batches with `slotNum < minSlotNum`.
          schema:
            $ref: '#/components/schemas/SlotNum'
        - name: maxSlothNum
          in: query
          required: false
          description: Only include batches with `slotNum > maxSlotNum`.
          schema:
            $ref: '#/components/schemas/SlotNum'
        - name: wonByEthereumAddress
          in: query
          required: false
          description: Only include slots won by a coordinator whose `forgerAddr == wonByEthereumAddress`.
          schema:
            $ref: '#/components/schemas/EthereumAddress'
        - name: finishedAuction
          in: query
          required: false
          description: If set to true, only include slots whose auction has finished.
          schema:
            type: boolean
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. Slots will be ordered by increasing `slotNum`.
          schema:
            type: string
            default: ASC
            enum:
                - ASC
                - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Slots'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
  '/state':
    get:
      tags:
        - Hermez status
      summary: Return information that represents the current state of the network.
      description: Return information that represents the current state of the network. It also includes metrics and statistics.
      operationId: getState
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/config':
    get:
      tags:
        - Hermez status
      summary: Return constant configuration of the network.
      description: Return constant configuration of the network.
      operationId: getConfig
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/tokens':
    get:
      tags:
        - Hermez status
      summary: Get information of the supported tokens in the Hermez network.
      description: Get information of the supported tokens in the Hermez network.
      operationId: getTokens
      parameters:
        - name: ids
          in: query
          required: false
          description: Include only specific tokens by their Hermez identifiers.
          schema:
            type: string
            description: Comma separated list of token identifiers
            example: "2,44,689"
        - name: symbols
          in: query
          required: false
          description: Include only specific tokens by their symbols.
          schema:
            type: string
            description: Comma separated list of token symbols.
            example: "DAI,NEC,UMA"
        - name: name
          in: query
          required: false
          description: Include token(s) by their names (or a substring of the name).
          schema:
            type: string
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. Tokens will be ordered by increasing tokenID.
          schema:
            type: string
            default: ASC
            enum:
                - ASC
                - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
components:
  schemas:
...
    HezEthereumAddress:
      type: string
      description: "Address of an Etherum account linked to the Hermez network."
      pattern: "^hez:0x[a-fA-F0-9]{40}$"
      example: "hez:0xaa942cfcd25ad4d90a62358b0dd84f33b398262a" 
    BJJ:
      type: string
      description: "BabyJubJub public key, encoded as base64 URL (RFC 4648), which result in 33 bytes. The padding byte is replaced by a sum of the encoded bytes."
      pattern: "^hez:[A-Za-z0-9_-]{44}$"
      example: "hez:rR7LXKal-av7I56Y0dEBCVmwc9zpoLY5ERhy5w7G-xwe"
...
    Accounts:
      type: object
      properties:
        accounts:
          type: array
          description: List of accounts.
          items: 
            $ref: '#/components/schemas/Account'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    Error400:
      allOf:
        - $ref: '#/components/schemas/Error'
        - example:
            message: Invalid signature.
    Error404:
      allOf:
        - $ref: '#/components/schemas/Error'
        - example:
            message: Item(s) not found.
...